New upstream release;
3.0.8
The changes in this release are as follows:
The http:authorization setting was added. SFTP was made to work with
FreeBSD servers. Zero length file transfer when modification time is
unknown was fixed. Uploading with the fish protocol when a file grows
was fixed. A core dump in fxp transfer mode was fixed.

3.0.9
The changes in this release are as follows:
This version fixes a hang when filtering output through an  external
program.