New upstream release;
3.0.13
The changes in this release are as follows:
The mirror option --no-symlinks has been fixed (sense was reversed). A
core dump has been fixed.

3.0.12
The changes in this release are as follows:
A new ftps:initial-prot setting was added. FTP over HTTP proxies with
the CONNECT method can now use proxy authorization. The "mirror"
command was fixed not to chmod unchanged files. Excluding files for
the sftp protocol was fixed. The transfer of growing files over the
fish protocol was fixed. mkdir over hftp was fixed. The delay between
connection attempts was fixed. The Polish translation was updated.