Upstream provides more packages than just those in my hardcoded list.
Bump revision.

Wow, it's hard to believe that I had a static list of binaries, copied
to %p/bin.  In the however-many years since I last cared to
(sincerely) update this package, things have changed.  I discovered
this when a colleague who also uses prover9/mace4 asked me to run some
utility, and I was dumbfounded when I found that it simply wasn't in
%p/bin.