Update midnight commander.
On a recent working trip I was inspired watching a colleague use mc so
effectively, so I thought I'd start to use it.  And, as a fink package
manager, I immediately set out to ensure that the mc I was using on my
machine was the absolute latest version.  It turned out that Max
Horn's package is slightly out of date, so I took a half hour or so to
upgrade it from 4.6.1 to 4.7.0.2 (the latest as of 2010-02-25).

It took some work.  Primary issue: the latest version of mc wants to
put stuff in %p/libexec.  I rerouted the various scripts that get
installed there by default to %p/bin.  That took a bit of hackery.